Add development tool to load memory measurement from file and apply it for simulation.
To use this, start the application without an eve online client process on the machine and drag a file containing the measurement on the  added textblock in the development tools section.
The bot then works with this loaded measurement.

To save a measurement to a file follow the guide at http://forum.botengine.de/thread/saving-memory-measurements-from-sanderling-to-file/